refactor: showErrorAlert 메서드 위치 수정 - UIViewController + extension 파일 생성 - delete 메서드 수정
